requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
scipy>=1.5.0
scikit-image>=0.17.0
tqdm>=4.46.0
open3d>=0.13.0
opencv-python>=4.5.0
timm>=0.4.5
